fix so that return value for plate and platecache is not always NodeList
The documentation says that plate and platecache can return a single
Node if there is only one model provided (as opposed to an array
of models). The fix makes it so that both Node and NodeList can be
returned as indicated by the docs.

The docs for plateText and platecacheText indicate that return value
is always an array, so the string and Node returing versions differ
in this respect. Will discuss changes if someone brings it up.